Repalettize and Reoptimize sprites
This is an automated fix for an issue discovered by GitHub user Shahelion (see https://github.com/malcolmriley/unused-textures/discussions/6).

Many thanks to Shahelion for reporting the issue and for their time and patience as I bumbled about seeking a solution.

As this is an automated fix, if you notice any visual degradation in the quality of the sprites, please let me know in a new discussion post.